+++ packages/terra-tabs/src/Tabs.jsx
@@ -1,17 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
-import classNamesBind from 'classnames/bind';
 import ThemeContext from 'terra-theme-context';
-import ContentContainer from 'terra-content-container';
-import ResponsiveElement from 'terra-responsive-element';
+import CommonTabItem from './common-tabs/CommonTabItem';
+import CommonTabs from './common-tabs/CommonTabs';
 import TabPane from './TabPane';
-import CollapsibleTabs from './_CollapsibleTabs';
-import CollapsedTabs from './_CollapsedTabs';
 import TabUtils from './TabUtils';
-import styles from './Tabs.module.scss';
-
-const cx = classNamesBind.bind(styles);
+import CommonTabContent from './common-tabs/CommonTabContent';
 
 /**
 NOTE: This is being commented out until discussions have been resolved around if modular tabs should be removed.
@@ -60,87 +54,27 @@ const propTypes = {
    * Key of the pane that should be open initially.
    */
   defaultActiveKey: PropTypes.string,
-
-  /**
-   * The viewport the element will be responsive to. One of `window`, `parent` or `none`.
-   * Note: If the responsive viewport is set to `none`, then tabs never _completely_
-   * collapse into a menu.
-   */
-  responsiveTo: PropTypes.oneOf(['window', 'parent', 'none']),
 };
 
 const defaultProps = {
   tabFill: false,
   fill: false,
-  responsiveTo: 'parent',
 };
 
 class Tabs extends React.Component {
   constructor(props) {
     super(props);
     this.getInitialState = this.getInitialState.bind(this);
-    this.getActiveTabIndex = this.getActiveTabIndex.bind(this);
-    this.handleOnChange = this.handleOnChange.bind(this);
-    this.handleTruncationChange = this.handleTruncationChange.bind(this);
-    this.wrapPaneOnClick = this.wrapPaneOnClick.bind(this);
-    this.setTabs = this.setTabs.bind(this);
     this.state = {
       activeKey: this.getInitialState(),
-      isLabelTruncated: false,
-      showCollapsedTabs: false,
     };
   }
 
   getInitialState() {
     if (this.props.activeKey) {
-      return null;
+      return this.props.activeKey;
     }
     return TabUtils.initialSelectedTabKey(this.props.children, this.props.defaultActiveKey);
-  }
-
-  handleOnChange(event, selectedPane) {
-    if (!selectedPane.props.isDisabled) {
-      if (this.props.onChange) {
-        this.props.onChange(event, selectedPane.key);
-      }
-      if (!this.props.activeKey && TabUtils.shouldHandleSelection(this.state.activeKey, selectedPane.key)) {
-        this.setState({ activeKey: selectedPane.key });
-      }
-    }
-  }
-
-  handleTruncationChange(isLabelTruncated) {
-    if (this.state.isLabelTruncated !== isLabelTruncated) {
-      this.setState({ isLabelTruncated });
-    }
-  }
-
-  setTabs(event) {
-    const showCollapsedTabs = event === 'tiny';
-    if (this.state.showCollapsedTabs !== showCollapsedTabs) {
-      this.setState({ showCollapsedTabs });
-    }
-  }
-
-  getActiveTabIndex() {
-    let activeIndex = -1;
-    React.Children.forEach(this.props.children, (child, index) => {
-      if (child.key === this.state.activeKey || child.key === this.props.activeKey) {
-        activeIndex = index;
-      }
-    });
-
-    return activeIndex;
-  }
-
-  wrapPaneOnClick(pane) {
-    return (event) => {
-      this.handleOnChange(event, pane);
-
-      if (pane.props.onClick) {
-        pane.props.onClick(event);
-      }
-    };
   }
 
   render() {
@@ -151,98 +85,44 @@ class Tabs extends React.Component {
       children,
       activeKey,
       defaultActiveKey,
-      responsiveTo,
       ...customProps
     } = this.props;
 
-    const theme = this.context;
+    const commonTabItems = [];
 
-    // NOTE: Hardcoding variant to structural until discussions have resolved around if we want modular tabs.
-    const variant = 'structural';
-    const tabsClassNames = classNames(cx(
-      'tabs-container',
-      { 'tab-fill': tabFill },
-      variant,
-      theme.className,
-    ),
-    customProps.className);
-
-    let content = null;
-    let isIconOnly = false;
-    const clonedPanes = [];
-    React.Children.forEach(children, (child, index) => {
-      let isActive = false;
-      if (child.key === this.state.activeKey || child.key === activeKey) {
-        isActive = true;
-        content = child.props.children;
+    React.Children.forEach(children, child => {
+      let content;
+      let tabContent;
+      if (child.key === this.state.activeKey) {
+        content = React.Children.map(child.props.children, contentItem => (
+          React.cloneElement(contentItem)
+        ));
+        tabContent = <CommonTabContent>{content}</CommonTabContent>;
       }
-
-      if (child.props.isIconOnly) {
-        isIconOnly = true;
-      }
-      clonedPanes.push(React.cloneElement(child, {
-        className: cx({ 'is-active': isActive }, child.props.className),
-        onClick: this.wrapPaneOnClick(child),
-        isActive,
-        index,
-      }));
+      commonTabItems.push(
+        <CommonTabItem
+          itemKey={child.key}
+          label={child.props.label}
+          icon={child.props.icon}
+          isIconOnly={child.props.isIconOnly}
+          showIcon={child.props.showIcon}
+          render={() => tabContent}
+          isDisabled={child.props.isDisabled}
+        />,
+      );
     });
 
-    content = React.Children.map(content, contentItem => (
-      React.cloneElement(contentItem, { isLabelHidden: isIconOnly || this.state.isLabelTruncated })
-    ));
-
-    const tabContent = (() => {
-      const collapsibleTabs = (
-        <CollapsibleTabs
-          activeKey={activeKey || this.state.activeKey}
-          activeIndex={this.getActiveTabIndex()}
-          onChange={this.handleOnChange}
-          onTruncationChange={this.handleTruncationChange}
-          variant={variant}
-        >
-          {clonedPanes}
-        </CollapsibleTabs>
-      );
-
-      if (responsiveTo === 'parent' || responsiveTo === 'window') {
-        const collapsedTabs = (
-          <CollapsedTabs
-            activeKey={activeKey || this.state.activeKey}
-            activeIndex={this.getActiveTabIndex()}
-            onTruncationChange={this.handleTruncationChange}
-          >
-            {clonedPanes}
-          </CollapsedTabs>
-        );
-
-        return (
-          <ResponsiveElement
-            onChange={this.setTabs}
-            responsiveTo={responsiveTo}
-          >
-            { this.state.showCollapsedTabs ? collapsedTabs : collapsibleTabs}
-          </ResponsiveElement>
-        );
-      }
-
-      return collapsibleTabs;
-    })();
-
     return (
-      <ContentContainer
+      <CommonTabs
+        id={customProps.id}
+        activeItemKey={this.state.activeKey}
+        onRequestActivate={key => this.setState({ activeKey: key })}
+        onChange={onChange}
+        variant="framework"
         {...customProps}
-        className={tabsClassNames}
-        fill={fill}
-        header={tabContent}
       >
-        <div
-          role="tabpanel"
-          className={cx('pane-content', { 'fill-parent': fill })}
-        >
-          {content}
-        </div>
-      </ContentContainer>
+        {commonTabItems}
+      </CommonTabs>
     );
   }
 }